+++ lib/src/authentication/has_api_tokens.dart
@@ -14,31 +14,53 @@ class HasApiTokens {
     return this;
   }
 
-  String createToken([
+  Map<String, dynamic> createToken([
     String guard = '',
     Duration? expiresIn,
+    bool withRefreshToken = false,
   ]) {
-    final jwt = JWT({'id': _userPayload?['id']});
-    String token = jwt.sign(SecretKey('${Config().get('key')}$guard'),
-        expiresIn: expiresIn ?? const Duration(hours: 24));
-    return token;
+    final jwt = JWT({'id': _userPayload?['id'], 'type': 'access_token'});
+    Map<String, dynamic> payload = {};
+    Duration expirationTime = expiresIn ?? const Duration(hours: 1);
+
+    String accessToken = jwt.sign(SecretKey('${Config().get('key')}$guard'),
+        expiresIn: expirationTime);
+
+    payload['access_token'] = accessToken;
+
+    if (withRefreshToken) {
+      final jwtRefresh =
+          JWT({'id': _userPayload?['id'], 'type': 'refresh_token'});
+      String refreshToken = jwtRefresh.sign(
+          SecretKey('${Config().get('key')}$guard'),
+          expiresIn: const Duration(days: 30));
+      payload['refresh_token'] = refreshToken;
+    }
+
+    payload['expires_in'] =
+        DateTime.now().add(expirationTime).toIso8601String();
+
+    return payload;
   }
 
-  String refreshToken(
+  Map<String, dynamic> refreshToken(
     String token, [
     String guard = '',
     Duration? expiresIn,
   ]) {
-    Map<String, dynamic> payload = JWT.decode(token).payload;
-    final jwt = JWT({'id': payload['id']});
-    String rTtoken = jwt.sign(SecretKey('${Config().get('key')}$guard'),
-        expiresIn: expiresIn ?? const Duration(hours: 24));
-    return rTtoken;
+    final jwt = verify(token, guard, 'refresh_token');
+    _userPayload = jwt;
+    return createToken(guard, expiresIn, true);
   }
 
-  Map<String, dynamic> verify(String token, String guard) {
+  Map<String, dynamic> verify(String token, String guard, String expectedType) {
     try {
       final jwt = JWT.verify(token, SecretKey('${Config().get('key')}$guard'));
+
+      if (jwt.payload['type'] != expectedType) {
+        throw Unauthenticated(message: 'Invalid token');
+      }
+
       return jwt.payload;
     } on JWTExpiredException {
       rethrow;