+++ lib/src/database/database_driver.dart
@@ -3,6 +3,7 @@ import 'package:vania/vania.dart';
 
 abstract class DatabaseDriver {
   const DatabaseDriver();
+  String get driver;
   Future<void> init([DatabaseConfig? config]);
   Connection get connection;
   Future<void> close();