+++ CHANGELOG.md
@@ -1,14 +1,26 @@
+## 0.0.3+1
+
+- Fixed bug: Authentication refresh token
+
+
+## 0.0.3+1
+
+- Fixed bug: migration columns length
+- Added sslmode to the MySqldriver
+
+## 0.0.3
+
+- Fixed Bug: Resolved issue with table creation in PostgreSQL
+
 ## 0.0.2+1
 
-- Adding bigIncrements column
-
+- Added bigIncrements and  softDeletes columns
 
 ## 0.0.2
 
-- Adding column index to vania file
-- Code format 
+- Added column index to vania file
+- Code formatted
 
- 
 ## 0.0.1
 
 - Initial version.